a lot of improvments
deleted all call to GOTO
modify rover.1
organized better functions

TODO delete file manager
