+++ rover.1
@@ -78,22 +78,22 @@ another tab that is pointed to the "destination" directory.
 .B ESC
 Quit rover.
 .TP
-.B arrow up/down
+.B Arrow up/down
 Move cursor up/down.
 .TP
 .B Page up/down
 Move cursor up/down 10 lines.
 .TP
-.B g/G
-Move cursor to top/bottom of listing.
+.B HOME/END
+Move cursor to top/bottom of the list.
 .TP
-.B l
+.B Arrow Right
 Enter selected directory.
 .TP
-.B h
+.B Arrow Left
 Go to parent directory.
 .TP
-.B H
+.B h
 Go to \fB$HOME\fR directory.
 .TP
 .B t
@@ -123,19 +123,19 @@ Open \fB$OPEN\fR with the selected file.
 .B /
 Start incremental search.
 .TP
-.B f/d/s
+.B F/D/H
 Toggle file/directory/hidden listing.
 .TP
 .B n/N
 Create new file/directory.
 .TP
-.B R
+.B F2
 Rename selected file or directory.
 .TP
 .B E
 Toggle execute permission of the selected file.
 .TP
-.B D
+.B CANC
 Delete selected file or (empty) directory.
 .TP
 .B m
