+++ src/config.h
@@ -18,7 +18,6 @@
 #include <locale.h> /* setlocale(), LC_ALL */
 #include <unistd.h> /* chdir(), getcwd(), read(), close(), ... */
 #include <dirent.h> /* DIR, struct dirent, opendir(), ... */
-#include <libgen.h>
 #include <sys/stat.h>
 #include <fcntl.h> /* open() */
 #include <sys/wait.h> /* waitpid() */
@@ -136,20 +135,6 @@ extern struct Rover rover;
 #define MAX(A, B) ((A) > (B) ? (A) : (B))
 #define ISDIR(E)  (strchr((E), '/') != NULL)
 
-/* Line Editing Macros. */
-#define EDIT_FULL(E)      ((E).left == (E).right)
-#define EDIT_CAN_LEFT(E)  ((E).left)
-#define EDIT_CAN_RIGHT(E) ((E).right < PATH_MAX - 1)
-#define EDIT_LEFT(E)      (E).buffer[(E).right--] = (E).buffer[--(E).left]
-#define EDIT_RIGHT(E)     (E).buffer[(E).left++] = (E).buffer[++(E).right]
-#define EDIT_INSERT(E, C) (E).buffer[(E).left++] = (C)
-#define EDIT_BACKSPACE(E) (E).left--
-#define EDIT_DELETE(E)    (E).right++
-#define EDIT_CLEAR(E)             \
-	{                             \
-		(E).left  = 0;            \
-		(E).right = PATH_MAX - 1; \
-	}
 /* Add / at the end of path */
 #define ADDSLASH(path)                     \
 	{                                      \
@@ -163,43 +148,22 @@ extern struct Rover rover;
 			free((p)); \
 		(p) = NULL;    \
 	}
-typedef enum EditStat {
-	CONTINUE,
-	CONFIRM,
-	CANCEL
-} EditStat;
 
 typedef int (*PROCESS)(const char *path);
 
 // FUnction declarations
-void init_marks(Marks *marks);
-void mark_none(Marks *marks);
-void add_mark(Marks *marks, char *dirpath, char *entry);
-void del_mark(Marks *marks, char *entry);
-void free_marks(Marks *marks);
-void reload();
+void reload(void);
 void sync_signals();
-int rover_getch();
-int rover_get_wch(wint_t *wch);
-void shell_escaped_cat(char *buf, char *str, size_t n);
 void update_view();
 int rowcmp(const void *a, const void *b);
 int ls(Row **rowsp, uint8_t flags);
 void free_rows(Row **rowsp, int nfiles);
 void cd(bool reset);
 void try_to_sel(const char *target);
-void reload();
 off_t count_dir(const char *path);
 off_t count_marked();
-int process_dir(PROCESS pre, PROCESS proc, PROCESS pos, const char *path);
-void process_marked(PROCESS pre, PROCESS proc, PROCESS pos, const char *msg_doing, const char *msg_done);
 void update_progress(off_t delta);
-int delfile(const char *path);
-int addfile(const char *path);
 int cpyfile(const char *srcpath);
-int adddir(const char *path);
 int movfile(const char *srcpath);
-void start_line_edit(const char *init_input);
-EditStat get_line_edit();
 
 #endif // _CONFIG_H
