+++ src/main.c
@@ -3,6 +3,27 @@
 
 struct Rover rover;
 
+static void init_marks(Marks *marks)
+{
+	strcpy(marks->dirpath, "");
+	marks->bulk     = BULK_INIT;
+	marks->nentries = 0;
+	marks->entries  = (char **)calloc(marks->bulk, sizeof(*marks->entries));
+}
+
+static void free_marks(Marks *marks)
+{
+	int i;
+
+	for (i = 0; marks->nentries && i < marks->bulk; i++)
+		if (marks->entries[i]) {
+			FREE(marks->entries[i]);
+			marks->nentries--;
+		}
+
+	FREE(marks->entries);
+}
+
 int main(int argc, char *argv[])
 {
 	int i;
